Add Js for selecting buttons
I'm trying to figure out how to select each button
store it in a variable to then compare it to a
number from 0-2 to determine the winner.

So far, I can select each button but for some
reason the "Rock" button generates 2 events.

Next steps: Figure out why, compare results to
produce a winner and then change the html element
to show who won.

It's hard so far but I'm having fun!